Calculate totalFee for orders
